requirements.txt
altair==4.1.0
astor==0.8.1
astroid==2.3.3
attrs==19.3.0
backcall==0.1.0
base58==2.0.0
bleach==3.1.4
blinker==1.4
boto3==1.12.40
botocore==1.15.40
cachetools==4.1.0
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.1
colorama==0.4.3
cycler==0.10.0
Cython==0.29.19
decorator==4.4.2
defusedxml==0.6.0
distython==0.0.3
docutils==0.15.2
entrypoints==0.3
enum-compat==0.0.3
future==0.18.2
gower==0.0.5
idna==2.9
ipykernel==5.2.1
ipython==7.13.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
isort==4.3.21
jedi==0.17.0
Jinja2==2.11.2
jmespath==0.9.5
joblib==0.14.1
jsonschema==3.2.0
jupyter-client==6.1.3
jupyter-core==4.6.3
kiwisolver==1.2.0
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
matplotlib==3.2.1
mccabe==0.6.1
mistune==0.8.4
nbconvert==5.6.1
nbformat==5.0.6
notebook==6.0.3
numpy==1.18.3
packaging==20.3
pandas==1.0.3
pandocfilters==1.4.2
parso==0.7.0
pathtools==0.1.2
pickleshare==0.7.5
Pillow==7.1.1
plotly==4.6.0
prometheus-client==0.7.1
prompt-toolkit==3.0.5
protobuf==3.11.3
pydeck==0.3.1
Pygments==2.6.1
pylint==2.4.4
pyparsing==2.4.7
pyrsistent==0.16.0
python-dateutil==2.8.1
pytz==2019.3
pywin32==227
pywinpty==0.5.7
pyzmq==19.0.0
requests==2.23.0
retrying==1.3.3
s3transfer==0.3.3
scikit-learn==0.21.3
scipy==1.4.1
seaborn==0.10.0
Send2Trash==1.5.0
six==1.14.0
sklearn==0.0
streamlit==0.58.0
terminado==0.8.3
testpath==0.4.4
toml==0.10.0
toolz==0.10.0
tornado==5.1.1
traitlets==4.3.3
tzlocal==2.0.0
unicodedata2==13.0.0.post2
Unidecode==1.1.1
urllib3==1.25.9
validators==0.14.3
watchdog==0.10.2
wcwidth==0.1.9
webencodings==0.5.1
widgetsnbextension==3.5.1
wincertstore==0.2
wordcloud==1.7.0
wrapt==1.11.2
